+++ services/gundeck/src/Gundeck/Client.hs
@@ -1,6 +1,5 @@
 module Gundeck.Client
-    ( register
-    , unregister
+    ( unregister
     , removeUser
     ) where
 
@@ -10,16 +9,11 @@ import Data.Id
 import Data.Predicate
 import Gundeck.Monad
 import Gundeck.Push.Native
-import Gundeck.Util
-import Network.Wai (Request, Response)
+import Network.Wai (Response)
 import Network.Wai.Utilities
 
 import qualified Gundeck.Notification.Data as Notifications
 import qualified Gundeck.Push.Data         as Push
-
--- | DEPRECATED: remove once brig is upgraded everywhere.
-register :: UserId ::: ClientId ::: Request ::: JSON ::: JSON -> Gundeck Response
-register (_uid ::: _cid ::: _req ::: _) = return empty
 
 unregister :: UserId ::: ClientId -> Gundeck Response
 unregister (uid ::: cid) = do